+++ private/buf/bufworkspace/workspace_provider.go
@@ -23,7 +23,10 @@ import (
 	"github.com/bufbuild/buf/private/buf/buftarget"
 	"github.com/bufbuild/buf/private/bufpkg/bufconfig"
 	"github.com/bufbuild/buf/private/bufpkg/bufmodule"
+	"github.com/bufbuild/buf/private/pkg/normalpath"
+	"github.com/bufbuild/buf/private/pkg/slicesext"
 	"github.com/bufbuild/buf/private/pkg/storage"
+	"github.com/bufbuild/buf/private/pkg/stringutil"
 	"github.com/bufbuild/buf/private/pkg/syserror"
 	"github.com/bufbuild/buf/private/pkg/tracing"
 	"github.com/gofrs/uuid/v5"
@@ -314,7 +317,10 @@ func (w *workspaceProvider) getWorkspaceForBucketAndModuleDirPathsV1Beta1OrV1(
 				return nil, err
 			}
 		}
-		moduleConfig, ok := v1WorkspaceTargeting.bucketIDToModuleConfig[moduleTargeting.moduleDirPath]
+		// Each moduleBucketAndTargeting represents a local module that we want to add to the moduleSet,
+		// and we look up its moduleConfig by its bucketID, because that is guaranteed to be unique (technically,
+		// moduleDirPath is also unique in v1/v1beta1, but just to be extra safe).
+		moduleConfig, ok := v1WorkspaceTargeting.bucketIDToModuleConfig[moduleBucketAndTargeting.bucketID]
 		if !ok {
 			// This should not happen since moduleBucketAndTargeting is derived from the module
 			// configs, however, we return this error as a safety check
@@ -322,7 +328,7 @@ func (w *workspaceProvider) getWorkspaceForBucketAndModuleDirPathsV1Beta1OrV1(
 		}
 		moduleSetBuilder.AddLocalModule(
 			mappedModuleBucket,
-			moduleTargeting.moduleDirPath, // bucket ID
+			moduleBucketAndTargeting.bucketID,
 			moduleTargeting.isTargetModule,
 			bufmodule.LocalModuleWithModuleFullName(moduleConfig.ModuleFullName()),
 			bufmodule.LocalModuleWithTargetPaths(
@@ -335,6 +341,13 @@ func (w *workspaceProvider) getWorkspaceForBucketAndModuleDirPathsV1Beta1OrV1(
 			),
 			bufmodule.LocalModuleWithV1Beta1OrV1BufYAMLObjectData(v1BufYAMLObjectData),
 			bufmodule.LocalModuleWithV1Beta1OrV1BufLockObjectData(v1BufLockObjectData),
+			bufmodule.LocalModuleWithDescription(
+				getLocalModuleDescription(
+					// See comments on getLocalModuleDescription.
+					moduleBucketAndTargeting.bucketID,
+					moduleConfig,
+				),
+			),
 		)
 	}
 	moduleSet, err := moduleSetBuilder.Build()
@@ -385,18 +398,41 @@ func (w *workspaceProvider) getWorkspaceForBucketBufYAMLV2(
 			)
 		}
 	}
+	// Only check for duplicate module description in v2, which would be an user error, i.e.
+	// This is not a system error:
+	// modules:
+	//   - path: proto
+	//     excludes:
+	//       - proot/foo
+	//   - path: proto
+	//     excludes:
+	//       - proot/foo
+	// but duplicate module description in v1 is a system error, which the ModuleSetBuilder catches.
+	seenModuleDescriptions := make(map[string]struct{})
 	for _, moduleBucketAndTargeting := range v2Targeting.moduleBucketsAndTargeting {
 		mappedModuleBucket := moduleBucketAndTargeting.bucket
 		moduleTargeting := moduleBucketAndTargeting.moduleTargeting
-		moduleConfig, ok := v2Targeting.bucketIDToModuleConfig[moduleTargeting.moduleDirPath]
+		// Each moduleBucketAndTargeting represents a local module that we want to add to the moduleSet,
+		// and we look up its moduleConfig by its bucketID, because that is guaranteed to be unique (moduleDirPaths
+		// are not in a v2 workspace).
+		moduleConfig, ok := v2Targeting.bucketIDToModuleConfig[moduleBucketAndTargeting.bucketID]
 		if !ok {
 			// This should not happen since moduleBucketAndTargeting is derived from the module
 			// configs, however, we return this error as a safety check
 			return nil, fmt.Errorf("no module config found for module at: %q", moduleTargeting.moduleDirPath)
 		}
+		moduleDescription := getLocalModuleDescription(
+			// See comments on getLocalModuleDescription.
+			moduleConfig.DirPath(),
+			moduleConfig,
+		)
+		if _, ok := seenModuleDescriptions[moduleDescription]; ok {
+			return nil, fmt.Errorf("multiple module configs found with the same description: %s", moduleDescription)
+		}
+		seenModuleDescriptions[moduleDescription] = struct{}{}
 		moduleSetBuilder.AddLocalModule(
 			mappedModuleBucket,
-			moduleTargeting.moduleDirPath,
+			moduleBucketAndTargeting.bucketID,
 			moduleTargeting.isTargetModule,
 			bufmodule.LocalModuleWithModuleFullName(moduleConfig.ModuleFullName()),
 			bufmodule.LocalModuleWithTargetPaths(
@@ -407,6 +443,7 @@ func (w *workspaceProvider) getWorkspaceForBucketBufYAMLV2(
 				moduleTargeting.moduleProtoFileTargetPath,
 				moduleTargeting.includePackageFiles,
 			),
+			bufmodule.LocalModuleWithDescription(moduleDescription),
 		)
 	}
 	moduleSet, err := moduleSetBuilder.Build()
@@ -455,4 +492,41 @@ func (w *workspaceProvider) getWorkspaceForBucketModuleSet(
 		configuredDepModuleRefs,
 		isV2,
 	), nil
+}
+
+// This formats a module name based on its module config entry in the v2 buf.yaml:
+// `path: foo, includes: ["foo/v1, "foo/v2"], excludes: "foo/v1/internal"`.
+//
+// For v1/v1beta1 modules, pathDescription should be bucketID.
+// For v2 modules, pathDescription should be moduleConfig.DirPath().
+//
+// We edit bucketIDs in v2 to include an index since directories can be overlapping.
+// We would want to use moduleConfig.DirPath() everywhere, but it is always "." in
+// v1/v1beta1, and it's not a good description.
+func getLocalModuleDescription(pathDescription string, moduleConfig bufconfig.ModuleConfig) string {
+	description := fmt.Sprintf("path: %q", pathDescription)
+	moduleDirPath := moduleConfig.DirPath()
+	relIncludePaths := moduleConfig.RootToIncludes()["."]
+	includePaths := slicesext.Map(relIncludePaths, func(relInclude string) string {
+		return normalpath.Join(moduleDirPath, relInclude)
+	})
+	switch len(includePaths) {
+	case 0:
+	case 1:
+		description = fmt.Sprintf("%s, includes: %q", description, includePaths[0])
+	default:
+		description = fmt.Sprintf("%s, includes: [%s]", description, stringutil.JoinSliceQuoted(includePaths, ", "))
+	}
+	relExcludePaths := moduleConfig.RootToExcludes()["."]
+	excludePaths := slicesext.Map(relExcludePaths, func(relInclude string) string {
+		return normalpath.Join(moduleDirPath, relInclude)
+	})
+	switch len(excludePaths) {
+	case 0:
+	case 1:
+		description = fmt.Sprintf("%s, excludes: %q", description, excludePaths[0])
+	default:
+		description = fmt.Sprintf("%s, excludes: [%s]", description, stringutil.JoinSliceQuoted(excludePaths, ", "))
+	}
+	return description
 }