+++ private/buf/cmd/buf/buf_test.go
@@ -1601,11 +1601,11 @@ buf.build/foo/bar
 		// format as json, sort by path
 		`
 {"path":"a_v1","name":"buf.build/foo/bar"}
-{"path":"b_no_name_v1","name":""}
+{"path":"b_no_name_v1"}
 {"path":"c_v1beta1","name":"buf.build/bar/baz"}
-{"path":"d_no_file","name":""}
-{"path":"e_no_file","name":""}
-{"path":"f_no_name_v1beta1","name":""}
+{"path":"d_no_file"}
+{"path":"e_no_file"}
+{"path":"f_no_name_v1beta1"}
 `,
 		"config",
 		"ls-modules",
@@ -1678,9 +1678,9 @@ buf.build/foo/bar
 		// format as json, sort by path
 		`
 {"path":"a","name":"buf.build/foo/bar"}
-{"path":"b_no_name","name":""}
+{"path":"b_no_name"}
 {"path":"c","name":"buf.build/bar/baz"}
-{"path":"d_no_name","name":""}
+{"path":"d_no_name"}
 `,
 		"config",
 		"ls-modules",
@@ -1779,7 +1779,7 @@ buf.build/foo/bar
 		t,
 		nil,
 		0,
-		`{"path":".","name":""}`,
+		`{"path":"."}`,
 		"config",
 		"ls-modules",
 		"--format",
@@ -1875,7 +1875,7 @@ func TestLsModulesModuleV1Beta1(t *testing.T) {
 		t,
 		nil,
 		0,
-		`{"path":".","name":""}`,
+		`{"path":"."}`,
 		"config",
 		"ls-modules",
 		"--format",
@@ -1930,7 +1930,7 @@ func TestLsModulesNoConfig(t *testing.T) {
 		t,
 		nil,
 		0,
-		`{"path":".","name":""}`,
+		`{"path":"."}`,
 		"config",
 		"ls-modules",
 		"--format",
@@ -2032,9 +2032,9 @@ func TestLsModulesConfigFlag(t *testing.T) {
 		0,
 		`
 {"path":"a","name":"buf.build/foo/bar"}
-{"path":"b_no_name","name":""}
+{"path":"b_no_name"}
 {"path":"c","name":"buf.build/bar/baz"}
-{"path":"d_no_name","name":""}
+{"path":"d_no_name"}
 `,
 		"config",
 		"ls-modules",
@@ -2045,7 +2045,7 @@ func TestLsModulesConfigFlag(t *testing.T) {
 	)
 }
 
-func TestLsModulesConfigPrecedence(t *testing.T) {
+func TestLsModulesConfigFlagTakesPrecedence(t *testing.T) {
 	// Cannot be parallel since we chdir.
 	pwd, err := osext.Getwd()
 	require.NoError(t, err)
@@ -2073,6 +2073,88 @@ d_no_name
 		"ls-modules",
 		"--config",
 		filepath.Join(pwd, "testdata", "lsmodules", "workspacev2", "buf.yaml"),
+	)
+}
+
+func TestLsModulesWorkspaceV2DuplicateDirPath(t *testing.T) {
+	// Cannot be parallel since we chdir.
+	pwd, err := osext.Getwd()
+	require.NoError(t, err)
+	defer func() {
+		r := recover()
+		assert.NoError(t, osext.Chdir(pwd))
+		if r != nil {
+			panic(r)
+		}
+	}()
+
+	require.NoError(t, osext.Chdir(filepath.Join(pwd, "testdata", "workspace", "success", "duplicate_dir_path")))
+	testRunStdout(
+		t,
+		nil,
+		0,
+		// default format is path
+		`
+proto/shared
+proto/shared
+proto/shared1
+proto/shared1
+separate
+`,
+		"config",
+		"ls-modules",
+	)
+	testRunStdout(
+		t,
+		nil,
+		0,
+		// format as path
+		`
+proto/shared
+proto/shared
+proto/shared1
+proto/shared1
+separate	
+	`,
+
+		"config",
+		"ls-modules",
+		"--format",
+		"path",
+	)
+	testRunStdout(
+		t,
+		nil,
+		0,
+		// format as name
+		`
+buf.build/shared/one
+buf.build/shared/zero
+	`,
+
+		"config",
+		"ls-modules",
+		"--format",
+		"name",
+	)
+	testRunStdout(
+		t,
+		nil,
+		0,
+		// format as json, sort by path
+		`
+	
+	{"path":"proto/shared","excludes":["proto/shared/prefix/foo"]}
+	{"path":"proto/shared","excludes":["proto/shared/prefix/bar"],"name":"buf.build/shared/zero"}
+	{"path":"proto/shared1","includes":["proto/shared1/prefix/x"],"name":"buf.build/shared/one"}
+	{"path":"proto/shared1","excludes":["proto/shared1/prefix/x"]}
+	{"path":"separate"}
+	`,
+
+		"config",
+		"ls-modules",
+		"--format",
+		"json",
 	)
 }