+++ private/buf/cmd/buf/command/config/configinit/configinit.go
@@ -135,13 +135,17 @@ func run(
 	moduleConfig, err := bufconfig.NewModuleConfig(
 		".",
 		moduleFullName,
+		// The default (empty) value for rootToIncludes and rootToExcludes only has key ".".
+		map[string][]string{
+			".": {},
+		},
 		map[string][]string{
 			".": {},
 		},
 		bufconfig.NewLintConfig(
 			bufconfig.NewEnabledCheckConfigForUseIDsAndCategories(
 				fileVersion,
-				[]string{"DEFAULT"},
+				[]string{"STANDARD"},
 			),
 			"",
 			false,