+++ private/bufpkg/bufcheck/multi_client.go
@@ -50,7 +50,7 @@ func (c *multiClient) Check(ctx context.Context, request check.Request) ([]*anno
 	requestRuleIDs := request.RuleIDs()
 	if len(requestRuleIDs) == 0 {
 		// If we didn't have specific ruleIDs, the requested ruleIDs are all default ruleIDs.
-		requestRuleIDs = slicesext.Map(slicesext.Filter(allRules, Rule.IsDefault), Rule.ID)
+		requestRuleIDs = slicesext.Map(slicesext.Filter(allRules, Rule.Default), Rule.ID)
 	}
 	// This is a map of the requested ruleIDs.
 	requestRuleIDMap := make(map[string]struct{})
@@ -97,6 +97,9 @@ func (c *multiClient) Check(ctx context.Context, request check.Request) ([]*anno
 			func(ctx context.Context) error {
 				delegateResponse, err := delegate.Client.Check(ctx, delegateRequest)
 				if err != nil {
+					if delegate.PluginName == "" {
+						return err
+					}
 					return fmt.Errorf("plugin %q failed: %w", delegate.PluginName, err)
 				}
 				lock.Lock()
@@ -151,6 +154,9 @@ func (c *multiClient) getRulesCategoriesAndChunkedIDs(ctx context.Context) (
 	for i, delegate := range c.checkClientSpecs {
 		delegateCheckRules, err := delegate.Client.ListRules(ctx)
 		if err != nil {
+			if delegate.PluginName == "" {
+				return nil, nil, nil, nil, err
+			}
 			return nil, nil, nil, nil, fmt.Errorf("plugin %q failed: %w", delegate.PluginName, err)
 		}
 		delegateRules := slicesext.Map(
@@ -167,6 +173,9 @@ func (c *multiClient) getRulesCategoriesAndChunkedIDs(ctx context.Context) (
 	for i, delegate := range c.checkClientSpecs {
 		delegateCheckCategories, err := delegate.Client.ListCategories(ctx)
 		if err != nil {
+			if delegate.PluginName == "" {
+				return nil, nil, nil, nil, err
+			}
 			return nil, nil, nil, nil, fmt.Errorf("plugin %q failed: %w", delegate.PluginName, err)
 		}
 		delegateCategories := slicesext.Map(