+++ private/bufpkg/bufmodule/bufmodule_test.go
@@ -377,7 +377,7 @@ func TestModuleCycleError(t *testing.T) {
 			"buf.build/foo/c",
 			"buf.build/foo/a",
 		},
-		moduleCycleError.OpaqueIDs,
+		moduleCycleError.Descriptions,
 	)
 
 	moduleB := moduleSet.GetModuleForOpaqueID("buf.build/foo/b")
@@ -394,7 +394,7 @@ func TestModuleCycleError(t *testing.T) {
 			"buf.build/foo/a",
 			"buf.build/foo/b",
 		},
-		moduleCycleError.OpaqueIDs,
+		moduleCycleError.Descriptions,
 	)
 
 	moduleC := moduleSet.GetModuleForOpaqueID("buf.build/foo/c")
@@ -411,7 +411,7 @@ func TestModuleCycleError(t *testing.T) {
 			"buf.build/foo/b",
 			"buf.build/foo/c",
 		},
-		moduleCycleError.OpaqueIDs,
+		moduleCycleError.Descriptions,
 	)
 }
 
@@ -461,7 +461,7 @@ func TestDuplicateProtoPathError(t *testing.T) {
 				"buf.build/foo/a",
 				"buf.build/foo/b",
 			},
-			duplicateProtoPathError.OpaqueIDs,
+			duplicateProtoPathError.ModuleDescriptions,
 		)
 	}
 	_, err = moduleA.ModuleDeps()
@@ -505,10 +505,10 @@ func TestNoProtoFilesError(t *testing.T) {
 		require.Error(t, err)
 		noProtoFilesError := &bufmodule.NoProtoFilesError{}
 		require.True(t, errors.As(err, &noProtoFilesError), err.Error())
-		require.Equal(
+		require.Contains(
 			t,
 			"buf.build/foo/b",
-			noProtoFilesError.OpaqueID,
+			noProtoFilesError.ModuleDescription,
 		)
 	}
 	moduleReadBucket := bufmodule.ModuleSetToModuleReadBucketWithOnlyProtoFiles(moduleSet)